Use unzip with overwrite to ease local development
During local development I repeatedly start and stop the docker container. What happens is that the downloaded graph cannot be unpacked as the files already exits. This leads to complete rebuild of the graph, thereby forcing me to wait before I can start the actual development task.

By using the -o flag the existing, identical graph is overwritten when I start the container the second time.